Fix crash on loading into a world

+++ src/main/java/net/lugo/utools/mixin/zoom/ZoomMixin.java
@@ -13,9 +13,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(GameRenderer.class)
 public class ZoomMixin {
     @Inject(at = @At("RETURN"), method = "getFov", cancellable = true)
-    public void getFov(Camera camera, float tickDelta, boolean changingFov, CallbackInfoReturnable<Double> callbackInfo) {
+    public void getFov(Camera camera, float tickDelta, boolean changingFov, CallbackInfoReturnable<Float> callbackInfo) {
         if (UTools.getConfig().turnOffZoom) return;
-        double fov = callbackInfo.getReturnValue();
+        float fov = callbackInfo.getReturnValue();
         float effectiveZoomMultiplier = Zoom.goal;
         Easing easing = Zoom.lastGoal >= Zoom.goal ? UTools.getConfig().zoomOutEasing : UTools.getConfig().zoomInEasing;
